Move PluginRenderer
